refactor(platform): extract the streaming UTF-8 decoder for cross-context reuse
take_decodable_utf8 was private to the agent terminal, but the shell PTY needs the
same carry-the-incomplete-tail behaviour. Moving it to platform keeps the shell out
of agent_runtime's internals rather than reaching across the context boundary.

Pure move: the function body and its two tests are unchanged, plus one added case
covering a genuine decode error so a malformed byte cannot stall the stream.
