+++ src-tauri/src/platform/mod.rs
@@ -11,3 +11,4 @@ pub(crate) mod logging;
 pub(crate) mod network;
 pub(crate) mod private_relay_fs;
 pub(crate) mod process;
+pub(crate) mod text;
